requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
crashtest==0.4.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
emoji==2.12.1
executing==2.0.1
fastapi==0.110.3
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==24.3.1
kubernetes==30.1.0
langchain==0.2.7
langchain-cli==0.0.25
langchain-community==0.2.7
langchain-core==0.2.29
langchain-huggingface==0.0.3
langchain-ollama==0.1.1
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langdetect==1.0.9
langserve==0.2.2
langsmith==0.1.85
libcst==1.4.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.1
onnxruntime==1.18.1
openai==1.35.13
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.15.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.2.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
# Editable install with no version control (rag_app==0.1.0)
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.1
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
trove-classifiers==2024.7.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.10
unstructured-client==0.24.1
urllib3==2.2.2
uvicorn==0.23.2
virtualenv==20.26.3
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
